Fix the failing test
Moved the testFairDataPointsFDP() test method from the LinkedDataServiceTest to the FAIRDataPointTest class in the fairdatapoint package.
