+++ backend/molgenis-emx2-fairdatapoint/src/test/java/org/molgenis/emx2/fairdatapoint/FAIRDataPointTest.java
@@ -1,21 +1,28 @@
-package org.molgenis.emx2.semantics.fairdatapoint;
+package org.molgenis.emx2.fairdatapoint;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.molgenis.emx2.fairdatapoint.FormatMimeTypes.formatToMediaType;
 
+import java.io.File;
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.nio.file.Path;
+import java.util.Objects;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.molgenis.emx2.Database;
 import org.molgenis.emx2.Schema;
+import org.molgenis.emx2.SchemaMetadata;
 import org.molgenis.emx2.datamodels.FAIRDataHubLoader;
-import org.molgenis.emx2.fairdatapoint.FAIRDataPoint;
-import org.molgenis.emx2.fairdatapoint.FAIRDataPointCatalog;
-import org.molgenis.emx2.fairdatapoint.FAIRDataPointDataset;
-import org.molgenis.emx2.fairdatapoint.FAIRDataPointDistribution;
+import org.molgenis.emx2.io.MolgenisIO;
+import org.molgenis.emx2.io.tablestore.TableStoreForXlsxFile;
+import org.molgenis.emx2.semantics.LinkedDataService;
 import org.molgenis.emx2.sql.TestDatabaseFactory;
+import org.molgenis.emx2.utils.StopWatch;
 import spark.Request;
 
 @Tag("slow")
@@ -23,10 +30,12 @@ public class FAIRDataPointTest {
 
   static Database database;
   static Schema[] fairDataHubSchemas;
+  static Schema fdpSchema;
 
   @BeforeAll
   public static void setup() {
     database = TestDatabaseFactory.getTestDatabase();
+    fdpSchema = database.dropCreateSchema("fdpTest");
     Schema fairDataHub_nr1 = database.dropCreateSchema("fairDataHub_nr1");
     Schema fairDataHub_nr2 = database.dropCreateSchema("fairDataHub_nr2 with a whitespace");
     FAIRDataHubLoader fairDataHubLoader = new FAIRDataHubLoader();
@@ -35,6 +44,29 @@ public class FAIRDataPointTest {
     fairDataHubSchemas = new Schema[2];
     fairDataHubSchemas[0] = fairDataHub_nr1;
     fairDataHubSchemas[1] = fairDataHub_nr2;
+  }
+
+  @Test
+  public void testFairDataPointsFDP() {
+    StopWatch.print("begin");
+
+    ClassLoader classLoader = getClass().getClassLoader();
+    Path file =
+        new File(Objects.requireNonNull(classLoader.getResource("fdp.xlsx")).getFile()).toPath();
+    MolgenisIO.importFromExcelFile(file, fdpSchema, true);
+
+    fdpSchema = database.getSchema("fdpTest");
+
+    // anyway, here goes the generation
+    StringWriter sw = new StringWriter();
+    LinkedDataService.getJsonLdForSchema(fdpSchema, new PrintWriter(sw));
+    System.out.println("result\r" + sw.getBuffer().toString());
+
+    assertEquals(2, fdpSchema.getTableNames().size());
+
+    sw = new StringWriter();
+    LinkedDataService.getTtlForSchema(fdpSchema, new PrintWriter(sw));
+    System.out.println(sw.toString());
   }
 
   @Test
@@ -182,5 +214,16 @@ public class FAIRDataPointTest {
         new FAIRDataPointDistribution(request, database);
     String result = fairDataPointDistribution.getResult();
     assertTrue(result.contains(formatToMediaType(format)));
+  }
+
+  private void runImportProcedure(TableStoreForXlsxFile store, SchemaMetadata cohortSchema) {
+    fdpSchema.migrate(cohortSchema);
+
+    StopWatch.print("creation of tables complete, now starting import data");
+
+    for (String tableName : fdpSchema.getTableNames()) {
+      if (store.containsTable(tableName))
+        fdpSchema.getTable(tableName).update(store.readTable(tableName)); // actually upsert
+    }
   }
 }
